requirements.txt
paho-mqtt
picamera2
torch
Pillow
numpy
ultralytics
RPi.GPIO
opencv-python